Rename 'external' to 'shared'
