+++ src/js/plugin/shared/PluginActions.js
