+++ src/js/plugin/shared/PluginDispatcher.js
