+++ src/js/plugin/shared/PluginEvents.js
