+++ src/js/plugin/shared/PluginHelper.js
